+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - Added: Optional PromiseKit support for sending.
-- Added: `connectAutomatically` parameter to connection creation. This will call `connect()` when requests are sent if not connected. 
+- Added: `connectAutomatically` parameter to connection creation. This will call `connect()` when requests are sent if not connected.
+- Added: `.getServices()` typed request.
 - Fixed: Calling `connect()` when already connected no longer disconnects and reconnects.
 
 ## [0.1.0] - 2021-03-05